requirements.txt
# Core dependencies
groq>=0.4.0
asyncio-throttle>=1.0.2
tenacity>=8.2.0
pydantic>=2.0.0
pyyaml>=6.0
click>=8.1.0

# Data processing
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0

# Database and analytics
dbt-core>=1.6.0
dbt-duckdb>=1.6.0
duckdb>=0.8.0

# Dashboard
streamlit>=1.28.0
plotly>=5.15.0
altair>=5.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
rich>=13.0.0
